Create example pom.xml maven
This code is an example that you can copy and insert into your pom.xml on Apache Netbeans. It is used to configure maven on Netbeans. So you can use mvn clean and install on terminal in your project.

Attention: Maven must be present on your PC and on the environment variables.

WSL: for those who use WSL just use apt install maven and you're done